feat: Add KMatcher interface and implementation
This commit adds the KMatcher interface and its implementation, kMatcher, to the kmatch package. The KMatcher interface defines a matcher that can be used to match strings based on certain criteria. The kMatcher struct implements the KMatcher interface and provides the Match method to determine if a given string matches the criteria defined in the KMatch struct.